Turn off flask-sqlalchemy modification tracking to silence a warning and save resources
We shouldn't need SQLALCHEMY_TRACK_MODIFICATIONS, so just disable it as suggested to
avoid the warning and save these little resources. This is a flask-sqlalchemy specific
session event tracking thing; sqlalchemy itself provides its own mechanism for this
these days, and if we need something like that, we should probably use those, not
this flask-sqlalchemy thing.